requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0
django-cors-middleware==1.5.0
django-environ==0.4.5
django-filter==2.2.0
django-graphql-jwt==0.3.1
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.9.1
graphene-subscriptions==1.0.2
graphql-core==2.3.1
graphql-relay==2.0.1
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Markdown==3.1.1
oauthlib==3.1.0
promise==2.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.7
whitenoise==5.0.1
zope.interface==5.1.0
